chore: move metrics related files together
